Switch to fractional/notional orders with per-position cap
Whole-share ordering is impractical on a $1k account — a single ASML
or VOO position would consume the entire balance. This switches to
Alpaca's notional (dollar-amount) buy orders so all symbols can
participate regardless of share price.

Key changes:
- executor.py: place_buy_order now takes notional (float) and uses
  MarketOrderRequest(notional=...) instead of qty=; place_sell_order
  takes shares: float for fractional qty on close
- risk.py: Position.shares is now float; compute_position_size returns
  float (no int() floor — fractional sizing is the point)
- config.py: adds MAX_POSITION_PCT (default 0.25) to cap any single
  position at 25% of equity regardless of risk-based sizing
- main.py: _execute_entry computes notional = min(risk-sized, cap),
  guards on notional < $1.0 instead of shares < 1
- database.py: shares column changed INTEGER → REAL
- docs/architecture.md: schema and .env example updated
- All 166 tests passing

Do NOT make these changes on the VPS until paper trading observation
completes — these were needed for Phase 4 live capital readiness.

+++ src/database.py
@@ -31,7 +31,7 @@ CREATE TABLE IF NOT EXISTS positions (
     symbol          TEXT    NOT NULL,
     entry_date      TEXT    NOT NULL,
     entry_price     REAL    NOT NULL,
-    shares          INTEGER NOT NULL,
+    shares          REAL    NOT NULL,
     stop_loss       REAL    NOT NULL,
     trailing_stop   REAL    NOT NULL,
     take_profit     REAL    NOT NULL,
@@ -111,7 +111,7 @@ def save_position(conn: sqlite3.Connection, position: Position) -> int:
     )
     conn.commit()
     db_id = cursor.lastrowid
-    logger.info("Saved position id=%d  %s  %d shares @ %.4f", db_id, position.symbol, position.shares, position.entry_price)
+    logger.info("Saved position id=%d  %s  %.4f shares @ %.4f", db_id, position.symbol, position.shares, position.entry_price)
     return db_id  # type: ignore[return-value]
